Add files via upload
This is the full pipeline for the processing of the full length videos, extracting the relevant  parts, extraction of the poses, keypoints (partial optional animation) and converting to Json file for the animator. Additionally mapping to the ASL gloss for the backend and database. 

Additionally, the excel file (the dataset) used in the extraction.