Add files via upload
The full notebook pipeline for downloading, processing the video, extracting the relevant timestamp of the video, processing the poses, keypoints, additional basic 2d animation, and conversion to Json file for the animator. All mapped to the ASL gloss for the benefit of the backend/database


additionally the excel file (the full dataset) used in the pipeline.